+++ clojure-mode.el
@@ -70,7 +70,6 @@
 (require 'align)
 (require 'subr-x)
 (require 'lisp-mnt)
-(require 'project)
 
 (declare-function lisp-fill-paragraph  "lisp-mode" (&optional justify))
 
@@ -558,9 +557,7 @@ replacement for `cljr-expand-let`."
                                        (beginning-of-line-text)
                                        (eq (get-text-property (point) 'face)
                                            'font-lock-doc-face)))
-                                'do-indent)))
-  ;; integration with project.el
-  (add-hook 'project-find-functions #'clojure-current-project))
+                                'do-indent))))
 
 (defcustom clojure-verify-major-mode t
   "If non-nil, warn when activating the wrong `major-mode'."
@@ -1733,16 +1730,6 @@ are cached in a buffer local variable (`clojure-cached-project-dir')."
       (setq clojure-cached-project-dir project-dir))
     project-dir))
 
-(defun clojure-current-project (&optional dir-name)
-  "Return the current project as a cons cell usable by project.el.
-
-Call is delegated down to `clojure-project-dir' with
-optional DIR-NAME as argument."
-  (let ((project-dir (clojure-project-dir dir-name)))
-    (if project-dir
-        (cons 'clojure project-dir)
-      nil)))
-
 (defun clojure-project-root-path (&optional dir-name)
   "Return the absolute path to the project's root directory.
 
@@ -1755,11 +1742,6 @@ Return nil if not inside a project."
                                 clojure-build-tool-files))))
     (when (> (length choices) 0)
       (car (sort choices #'file-in-directory-p)))))
-
-;; project.el integration
-(cl-defmethod project-roots ((project (head clojure)))
-  "Return the list of directory roots of the PROJECT."
-  (list (cdr project)))
 
 (defun clojure-project-relative-path (path)
   "Denormalize PATH by making it relative to the project root."